+++ app/__init__.py
@@ -1,3 +1,3 @@
 """Telegram Zammad Gateway."""
 
-__version__ = "0.2.0"
+__version__ = "0.3.0"